+++ app/src/main/java/com/teester/whatsnearby/data/location/LocationJobPresenter.java
@@ -11,14 +11,11 @@ import com.mapzen.android.lost.api.LocationListener;
 import com.mapzen.android.lost.api.LocationRequest;
 import com.mapzen.android.lost.api.LocationServices;
 import com.mapzen.android.lost.api.LostApiClient;
-import com.teester.whatsnearby.BuildConfig;
 import com.teester.whatsnearby.data.PreferenceList;
 import com.teester.whatsnearby.data.source.Preferences;
 import com.teester.whatsnearby.data.source.QueryOverpass;
 import com.teester.whatsnearby.data.source.SourceContract;
 import com.teester.whatsnearby.main.MainActivity;
-
-import java.util.Locale;
 
 import static android.support.v4.content.ContextCompat.startActivity;
 
@@ -29,9 +26,6 @@ public class LocationJobPresenter
 		LocationListener,
 		Runnable {
 
-	private static final int MINQUERYINTERVAL = 60 * 60 * 1000;
-	private static final double MINQUERYDISTANCE = 20;
-
 	private Location location;
 	private Location lastLocation;
 	private Location lastQueryLocation;
@@ -39,8 +33,6 @@ public class LocationJobPresenter
 	private LostApiClient client;
 	private Context context;
 	private SourceContract.Preferences preferences;
-	private boolean query;
-	private String notQueryReason;
 
 	public LocationJobPresenter(Context context, LocationContract.LocationJobService locationCallback) {
 		this.context = context;
@@ -72,7 +64,8 @@ public class LocationJobPresenter
 			lastQueryLocation = setPreviousLocation(PreferenceList.LAST_QUERY_LOCATION_LATITUDE, PreferenceList.LAST_QUERY_LOCATION_LONGITUDE);
 		}
 
-		decideWhetherToQuery();
+		QueryOrNot decision = new QueryOrNot(preferences, location);
+		boolean query = decision.getWhetherToQuery();
 
 		if (query) {
 			new Thread(this).start();
@@ -120,87 +113,6 @@ public class LocationJobPresenter
 		if (queried) {
 			preferences.setDoublePreference(PreferenceList.LAST_QUERY_LOCATION_LATITUDE, location.getLatitude());
 			preferences.setDoublePreference(PreferenceList.LAST_QUERY_LOCATION_LONGITUDE, location.getLongitude());
-		}
-	}
-
-	/**
-	 * Logic dictating whether or not to query the Overpass api for a given location based on
-	 * location accuracy, time since last notification and distance since last query.  If the app is in
-	 * debug mode, it always returns true
-	 */
-	private void decideWhetherToQuery() {
-		query = true;
-		notQueryReason = "";
-
-		boolean debug_mode = preferences.getBooleanPreference(PreferenceList.DEBUG_MODE);
-
-        checkTimeAndDistanceSinceLastLocation();
-		checkNumberOfDetections();
-
-		// If we're in debug mode, query every time
-		if (debug_mode && BuildConfig.DEBUG) {
-			query = true;
-		}
-
-		if (query) {
-			notQueryReason += "• Queried\n";
-		}
-
-		preferences.setStringPreference(PreferenceList.NOT_QUERY_REASON, notQueryReason);
-	}
-
-	/**
-     * Don't query Overpass if less than 1 hour has passed since the last notification
-     *
-     * Don't query Overpass is you've moved more than 20m from the last location query (5 mins ago)
-     * (indicates you're probably not in the same place as 5 mins ago)
-     *
-     * Don't query Overpass is youre still within 20m of the last location query that you were
-     * notified about (indicates you've probably still in the same place)
-	 */
-    private void checkTimeAndDistanceSinceLastLocation() {
-		long lastNotificationTime = preferences.getLongPreference(PreferenceList.LAST_NOTIFICATION_TIME);
-
-        if (System.currentTimeMillis() - lastNotificationTime < MINQUERYINTERVAL) {
-			notQueryReason += String.format(Locale.getDefault(), "• Not long enough since last notification: %dmins\n", ((System.currentTimeMillis() - lastNotificationTime) / 60000));
-			query = false;
-		}
-		if (location.distanceTo(lastLocation) > MINQUERYDISTANCE) {
-			notQueryReason += String.format(Locale.getDefault(), "• Too far from last location: %.0fm\n", location.distanceTo(lastLocation));
-			query = false;
-		}
-		if (location.distanceTo(lastQueryLocation) < MINQUERYDISTANCE) {
-			notQueryReason += String.format(Locale.getDefault(), "• Not far enough from location of last query: %.0fm\n", location.distanceTo(lastQueryLocation));
-			query = false;
-		}
-	}
-
-	/**
-	 * Query based on the number of location detections close to this location in a row.
-	 * Decide how many detections are needed based on accuracy
-	 */
-	private void checkNumberOfDetections() {
-
-		long detections = preferences.getLongPreference(PreferenceList.NUMBER_OF_VISITS);
-		float accuracy = location.getAccuracy();
-		int numberOfDetectionsRequired = 2;
-
-		if (accuracy > 100) {
-			numberOfDetectionsRequired = 3;
-		}
-		if (accuracy > 1000) {
-			numberOfDetectionsRequired = 4;
-		}
-
-		if (location.distanceTo(lastLocation) < 20) {
-			preferences.setLongPreference(PreferenceList.NUMBER_OF_VISITS, detections + 1);
-		} else {
-			preferences.setLongPreference(PreferenceList.NUMBER_OF_VISITS, 1);
-		}
-
-		if (detections < numberOfDetectionsRequired) {
-			query = false;
-			notQueryReason += "• Not enough detections in a row";
 		}
 	}
